==> 5_Chd7_KO_RNA/codes/2_0_DEseq2_add_matchGN.ipynb <==
library(tidyverse)
library(readr)
library(GenomicFeatures)
library(DESeq2)
library(org.Mm.eg.db)
library(rjson)
library(tximport)
library(DBI)
library(rje)
library(plyr)

codedir <- getwd()
outdir <- '/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/0_salmon_DE_out/2_DEseq/ENSEMBL_ID'
outdir.compiled <- '/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/0_salmon_DE_out/2_DEseq_compiled'
gndir <- '/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/0_salmon_DE_out/2_DEseq/gene_name'
outdir.compiled <- '/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/0_salmon_DE_out/2_DEseq_compiled'

######################################## Convert ENSEMBL ID to gene symbols ########################################
# Download convert table from:http://useast.ensembl.org/biomart/martview/8c1957c27101a044a318d51140a289e1

cv_file <- '/home/pipkin/references/mm_BioMart_GeneStableID_GeneName.txt'
cv_tb <- read_csv(cv_file)

matchGN <- function(input, outfilename, cvTb=cv_tb){
    colnames(input) <- c("ensembl_stable_ID", colnames(input)[2:length(colnames(input))])
    output <- cvTb %>% right_join(input, by="ensembl_stable_ID")
    output$ensembl_stable_ID <- NULL
    write_csv(output, outfilename)
}

## 0. Prepare reference

###--- Make reference
#txdb <- makeTxDbFromGFF('/home/pipkin/references/GRCm38.99/Mus_musculus.GRCm38.99.gtf')
#saveDb(txdb, file='/home/pipkin/references/GRCm38.99/Mus_musculus.GRCm38.99')
mmRef <- '/home/pipkin/references/GRCm38.102/Mus_musculus.GRCm38.102'

###--- Convert transcript ID to gene ID
txdb <- loadDb(mmRef)
k <- keys(txdb, "GENEID")
res <- AnnotationDbi::select(txdb, k, "TXNAME", "GENEID")
tx2gene <- res[,2:1]
#write_csv(tx2gene, '/home/pipkin/references/GRCm38.102/tx2gene.csv')

##########---------- Read Quant Files
colData <- read.csv('/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/Info/meta.csv')
files <- file.path("/media/pipkin/ROCKET-PRO/CD8_DEV_SC/5_Chd7_KO_RNA/0_salmon_DE_out/1_salmon_nonTrim",
                   colData$Name,"quant.sf")
names(files) <- colData$Cond
txi <- tximport(files, type="salmon", tx2gene=tx2gene, ignoreTxVersion = TRUE, dropInfReps = TRUE) # Drop in freps TURE = ignore verison  # Ignore TX verison stringsplits on . 

# Construct sampleTable
sampleTable <- data.frame(condition =factor(colData$Type))
rownames(sampleTable) <- colnames(txi$counts)

#import into DESEQ2 framework
dds <- DESeqDataSetFromTximport(txi, sampleTable, ~ condition)
summary(dds)


dds <- DESeq(dds) #RunDESEQ

norm.df <- data.frame(counts(dds, normalized=T))
colnames(norm.df) <- colData$Name
norm.name <- file.path(outdir.compiled, "DESeq2_normalized_counts_bysample.csv")
norm.name.gn <- file.path(outdir.compiled, "DESeq2_normalized_counts_bysample_gn.csv")
write.csv(norm.df, norm.name)

matchGN(read_csv(norm.name), norm.name.gn)

for (i in c("Naive", "48h", "EE", "TE", "MP", "DP")){
    KO_i <- paste("KO", i, sep="_")
    WT_i <- paste("WT", i, sep="_")
    print(KO_i)
    print(WT_i)
    results <- as.data.frame(results(dds, contrast = c("condition",WT_i,KO_i)))
    write.csv(results, file.path(outdir, paste(WT_i, '-', KO_i, "-Salmon.csv", sep="")))
    results <- as.data.frame(results(dds, contrast = c("condition",KO_i,WT_i)))
    write.csv(results, file.path(outdir, paste(KO_i, '-', WT_i, "-Salmon.csv", sep="")))
}

for (i in c("Naive", "48h", "EE", "TE", "MP", "DP")){
    for (j in c("Naive", "48h", "EE", "TE", "MP", "DP")) {
        if (i != j){
            KO_i <- paste("KO", i, sep="_")
            KO_j <- paste("KO", j, sep="_")
            WT_i <- paste("WT", i, sep="_")
            WT_j <- paste("WT", j, sep="_")
            print(KO_i)
            print(KO_j)
            results <- as.data.frame(results(dds, contrast = c("condition",KO_i,KO_j)))
            write.csv(results, file.path(outdir, paste(KO_i, '-', KO_j, "-Salmon.csv", sep="")))
            results <- as.data.frame(results(dds, contrast = c("condition",WT_i,WT_j)))
            write.csv(results, file.path(outdir, paste(WT_i, '-', WT_j, "-Salmon.csv", sep="")))
        }
    }
}

for (de_out in list.files(path=outdir, pattern='.csv')){
    de_out_file <- file.path(outdir, de_out)
    new_out_file <- file.path(gndir, gsub(".csv", "_gn.csv", de_out))
    matchGN(read_csv(de_out_file), new_out_file)
}
